recognise/requirements.txt
python-dotenv
shazamio
spotipy
